Move roster view into homerooms controller, nested routes not necessary here
