+++ app/controllers/students_controller.rb
@@ -1,9 +1,6 @@
-require 'csv'
-
 class StudentsController < ApplicationController
 
   before_action :authenticate_educator!
-  before_action :assign_homeroom
 
   def show
     @student = Student.find(params[:id])
@@ -15,7 +12,7 @@ class StudentsController < ApplicationController
     @intervention = Intervention.new
     @interventions = @student.interventions
 
-    @roster_url = homeroom_students_path(@student.homeroom)
+    @roster_url = homeroom_path(@student.homeroom)
     @csv_url = student_path(@student) + ".csv"
     @student_url = student_path(@student)
 
@@ -26,22 +23,4 @@ class StudentsController < ApplicationController
     end
   end
 
-  def index
-    @students = @homeroom.students
-    # Order for dropdown menu of homerooms
-    @homerooms_by_name = Homeroom.where.not(name: "Demo").order(:name)
-    @current_school_year = SchoolYear.date_to_school_year(Time.new)
-  end
-
-  private
-
-  def assign_homeroom
-    @homeroom = Homeroom.friendly.find(params[:homeroom_id])
-  rescue ActiveRecord::RecordNotFound
-    if current_educator.homeroom.present?
-      @homeroom = current_educator.homeroom
-    else
-      @homeroom = Homeroom.first
-    end
-  end
 end
