+++ app/views/homerooms/_map_row_fragment.html.erb
