+++ app/views/homerooms/_mcas_row_fragment.html.erb
