+++ app/views/homerooms/show.html.erb
@@ -97,9 +97,9 @@
      </thead>
       <% if @students.present? %>
         <!-- ROWS -->
-        <% @students.find_each do |student| %>
+        <% @students.each do |student| %>
           <% row_data = StudentRowData.new(student, @current_school_year) %>
-          <%= render "students/student_row", student: student, row_data: row_data, presenter: StudentPresenter.new(student), risk: row_data.risk unless student.nil? %>
+          <%= render "student_row", student: student, row_data: row_data, presenter: StudentPresenter.new(student), risk: row_data.risk unless student.nil? %>
         <% end %>
       <% end %>
     </table>
